% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cetesR-package.R
\docType{package}
\name{cetesR-package}
\alias{cetesR}
\alias{cetesR-package}
\title{cetesR: Herramientas para Valores Gubernamentales de México}
\description{
El paquete \code{cetesR} proporciona funciones para consultar, procesar y
analizar información financiera de los valores gubernamentales de México
(CETES, Bonos M y Udibonos). Extrae datos directamente desde la API oficial
del Banco de México (Banxico) y complementa las series con cotizaciones
del mercado secundario.
}
\details{
Las funciones principales incluidas en el paquete son:
\itemize{
\item \code{\link{getBono}}: Función principal para obtener el historial consolidado, plazos, montos y tasas de un bono.
\item \code{\link{set_banxico_token}}: Administrador para configurar el token de la API de Banxico.
\item \code{\link{consultaApiDatos}}: Consulta directa de series históricas en la API de Banxico.
\item \code{\link{consultaApiDatoUltimo}}: Consulta del dato más reciente publicado para una serie.
\item \code{\link{consultaApiMeta}}: Extracción de metadatos de las series de Banxico.
\item \code{\link{catalogo_series}}: Dataset interno con la tabla de mapeo de identificadores y series.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abemorc/cetesR}
  \item \url{https://abemorc.github.io/cetesR/}
  \item Report bugs at \url{https://github.com/abemorc/cetesR/issues}
}

}
\author{
\strong{Maintainer}: Abelardo Morales Calva \email{abemorca@gmail.com}

Authors:
\itemize{
  \item Abelardo Morales Calva \email{abemorca@gmail.com}
}

}
\keyword{internal}
