% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultaApiDatos.R
\name{consultaApiDatos}
\alias{consultaApiDatos}
\title{Consulta de Datos Históricos en la API de Banxico}
\usage{
consultaApiDatos(
  codigo,
  fechaInicial = NULL,
  fechaFinal = NULL,
  token = Sys.getenv("BANXICO_TOKEN"),
  idioma = "es"
)
}
\arguments{
\item{codigo}{\code{character} Código exacto de la serie de tiempo a consultar.
Puede buscar el código correspondiente a cada instrumento mediante el dataset interno \code{catalogo_series}.}

\item{fechaInicial}{\code{character} o \code{Date}. Fecha a partir de la cual se quiere consultar los datos (formato "YYYY-MM-DD").}

\item{fechaFinal}{\code{character} o \code{Date}. Fecha hasta la cual se quiere consultar los datos. Se recomienda usar \code{Sys.Date()} para la fecha actual.}

\item{token}{\code{character}. Token de autenticación proporcionado por Banxico.
Por defecto, lo busca en la variable de entorno \code{BANXICO_TOKEN}.
Se recomienda configurarlo previamente con \code{set_banxico_token()}.}

\item{idioma}{\code{character} Idioma para la consulta de los metadatos devueltos por el servidor.
\itemize{
\item "es" Español (por defecto).
\item "en" Inglés.
}}
}
\value{
Un \code{data.frame} estructurado con dos columnas:
\itemize{
\item \strong{Fecha}: Clase \code{Date}.
\item \strong{Código de la serie}: Clase \code{numeric} con los valores correspondientes.
}
}
\description{
Extrae la serie de tiempo histórica de un instrumento financiero directamente
desde el Sistema de Información Económica (SIE) del Banco de México (Banxico).
}
\details{
Esta función realiza una petición HTTP tipo GET a la API REST de Banxico.
Procesa la respuesta en formato JSON, estructura los datos y realiza la
conversión automática de tipos (las fechas se transforman a clase \code{Date}
y los valores a clase \code{numeric}).

En caso de que el servidor devuelva un error (como un token inválido, exceder
el límite de consultas o solicitar una serie inexistente), la función captura
y muestra el mensaje oficial detallado de Banxico.
}
\examples{
\dontrun{
# 1. Configurar el token previamente
# set_banxico_token("TU_TOKEN_AQUI")

# 2. Consultar la serie histórica (ej. Tasa de rendimiento CETES 28)
datos_cetes <- consultaApiDatos(
  codigo = "SF43936",
  fechaInicial = '2020-01-01',
  fechaFinal = Sys.Date()
)

head(datos_cetes)
}
}
