% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consutaApiMeta.R
\name{consultaApiMeta}
\alias{consultaApiMeta}
\title{Consulta de Metadatos en la API de Banxico}
\usage{
consultaApiMeta(codigo, token = Sys.getenv("BANXICO_TOKEN"), idioma = "es")
}
\arguments{
\item{codigo}{\code{character} Código exacto de la serie de tiempo a consultar.
Puede buscar el código correspondiente a cada instrumento mediante el dataset interno \code{catalogo_series}.}

\item{token}{\code{character}. Token de autenticación proporcionado por Banxico.
Por defecto, lo busca en la variable de entorno \code{BANXICO_TOKEN}.
Se recomienda configurarlo previamente con \code{set_banxico_token()}.}

\item{idioma}{\code{character} Idioma para la consulta de los metadatos devueltos por el servidor.
\itemize{
\item "es" Español (por defecto).
\item "en" Inglés.
}}
}
\value{
Un \code{data.frame} de una sola fila con las siguientes columnas:
\itemize{
\item \strong{idSerie}: Código identificador de la serie.
\item \strong{titulo}: Nombre descriptivo de la serie.
\item \strong{fechaInicio}: Clase \code{Date}. Fecha inicial desde la que se tiene registro.
\item \strong{fechaFin}: Clase \code{Date}. Fecha de la última actualización disponible.
\item \strong{periodicidad}: Frecuencia con la que se actualiza la información (ej. Semanal).
\item \strong{cifra}: Tipo de cifra (ej. Oportuna, Revisada).
\item \strong{unidad}: Unidad de medida con la que se expresa la serie.
\item \strong{versionada}: \code{logical}. \code{TRUE} o \code{FALSE} si la serie tiene múltiples versiones.
}
}
\description{
Extrae la información descriptiva (metadatos) asociada a una serie de tiempo
específica desde el Sistema de Información Económica (SIE) de Banxico.
}
\details{
Esta función es útil para conocer el contexto de la información antes de
descargar los datos históricos. Permite verificar la fecha en la que inicia
el registro, la periodicidad de actualización y la unidad de medida en la
que está expresada la serie (ej. porcentajes, millones de pesos).
}
\examples{
\dontrun{
# 1. Asegúrate de tener configurado tu token
# set_banxico_token("TU_TOKEN_AQUI")

# 2. Consultar los metadatos de la Tasa de rendimiento de CETES 28
metadatos <- consultaApiMeta(codigo = "SF43936")

print(metadatos)
}
}
