% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cetesR-package.R
\docType{package}
\name{cetesR-package}
\alias{cetesR}
\alias{cetesR-package}
\title{cetesR: Reunir y analizar informacion de los valores gubernamentales en México}
\description{
Obtener informacion de banxico y diversos instrumentos financieros.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abemorc/cetesR}
  \item \url{https://abemorc.github.io/cetesR/}
  \item Report bugs at \url{https://github.com/abemorc/cetesR/issues}
}

}
\author{
\strong{Maintainer}: Abelardo Morales \email{abemorca@gmail.com}

Authors:
\itemize{
  \item Martin Medina \email{martin_medina0308@hotmail.com}
}

}
\keyword{internal}
