% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cetesR-package.R
\docType{package}
\name{cetesR-package}
\alias{cetesR}
\alias{cetesR-package}
\title{cetesR: Reunir y analizar informacion de los valores gubernamentales en México}
\description{
Obtener informacion de banxico y diversos brokers para la valuacion y analisis de valores gubernamentales de M¿éxico.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abemorc/cetesR}
  \item Report bugs at \url{https://github.com/abemorc/cetesR/issues}
}

}
\author{
\strong{Maintainer}: Abelardo Morales Calva \email{abemorca@gmail.com}

}
\keyword{internal}
