% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultaApiDatos.R
\name{consultaAPiDatos}
\alias{consultaAPiDatos}
\title{Consulta de datos en la API de banxico}
\usage{
consultaAPiDatos(
  codigo,
  fechaInicial = NULL,
  fechaFinal = NULL,
  token,
  idioma = "es"
)
}
\arguments{
\item{codigo}{Código de la serie de tiempo a consultar, puede consultar el
código mediante \code{\link[=catalogo_series]{catalogo_series()}}.}

\item{fechaInicial}{Fecha a partir de la cual se quiere consultar los datos.}

\item{fechaFinal}{Fecha hasta la cual se quiere consultar los datos, puede
usar \code{\link[=Sys.Date]{Sys.Date()}} para que la fecha de la consulta sea hasta la fecha
actual}

\item{token}{Token proporcionado por Banxico para la consulta. Puede
obtenerlo en la pagina de
\href{https://www.banxico.org.mx/SieAPIRest/service/v1/token}{Banxico}}

\item{idioma}{\code{character}  Idioma para la consulta de los metadatos.
\itemize{
\item es Español
\item en Inglés
}}
}
\value{
\code{\link[=data.frame]{data.frame()}} con las siguientes columnas:
\itemize{
\item Fecha
\item Valor
}
}
\description{
Consulta de datos en la API de banxico
}
\examples{
\dontrun{

token <- "tu token"
datos <- consultaAPi("SF43936", '2015-01-01', Sys.Date(), token = token)
}
}
