% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultaApiDatoUltimo.R
\name{consultaApiDatoUltimo}
\alias{consultaApiDatoUltimo}
\title{Consulta del valor mas reciente de la serie de datos}
\usage{
consultaApiDatoUltimo(codigo, token)
}
\arguments{
\item{codigo}{Código de la serie de tiempo a consultar, puede consultar el
código mediante \code{\link[=catalogo_series]{catalogo_series()}}.}

\item{token}{Token proporcionado por Banxico para la consulta. Puede
obtenerlo en la pagina de
\href{https://www.banxico.org.mx/SieAPIRest/service/v1/token}{Banxico}}
}
\value{
\code{\link[=data.frame]{data.frame()}} con las siguientes columnas:
\itemize{
\item \strong{Fecha}
\item \strong{Código de la serie}
}
}
\description{
Consulta del valor mas reciente de la serie de datos
}
\examples{
token <- "fb88d18e41bebb656375cd9b4db5878253b564b3742935ca96c9bd9fb67e1274"
datos <- consultaApiDatoUltimo("SF43936", token = token)

}
