% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consutaApiMeta.R
\name{consultaApiMeta}
\alias{consultaApiMeta}
\title{Consulta de metadatos en la API de banxico}
\usage{
consultaApiMeta(codigo, token, idioma = "es")
}
\arguments{
\item{codigo}{Código de la serie de tiempo a consultar, puede consultar el
código mediante \code{\link[=catalogo_series]{catalogo_series()}}.}

\item{token}{Token proporcionado por Banxico para la consulta. Puede
obtenerlo en la pagina de
\href{https://www.banxico.org.mx/SieAPIRest/service/v1/token}{Banxico}}

\item{idioma}{\code{character}  Idioma para la consulta de los metadatos.
\itemize{
\item "es" Español
\item "en" Inglés
}}
}
\value{
\code{\link[=data.frame]{data.frame()}} con las siguientes columnas:
\itemize{
\item idSerie Código de la serie
\item titulo Nombre de la serie
\item fechaInicio Fecha inicial desde la que se tiene registro
\item fechaFin Fecha hasta la que se tiene registro
\item periodicidad Frecuencia con la que se actualiza la información
\item cifra .
\item unidad Unidad de medida con la que se expresa la serie
\item versionada \code{TRUE} o \code{FALSE}
}
}
\description{
Consulta de metadatos en la API de banxico
}
\examples{
token <- "fb88d18e41bebb656375cd9b4db5878253b564b3742935ca96c9bd9fb67e1274"
metadatos <- consultaApiMeta("SF43936", token = token)

}
