% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBono.R
\name{getBono}
\alias{getBono}
\title{Información Bono Gubernamental}
\usage{
getBono(bono, fechaInicial, fechaFinal, token, idioma = "es")
}
\arguments{
\item{bono}{\code{characther}  Nombre del bono gubernamental del que se va a
obtener la información, puede ser uno de los siguientes: c("cetes28",
"cetes91", "cetes182", "cetes364", "cetes728", "bonosM3", "bonosM5",
"bonosM7", "bonosM10", "bonosM20", "bonosM30", "udibonos3", "udibonos5",
"udibonos10", "udibonos20", "udibonos30")}

\item{fechaInicial}{Fecha a partir de la cual se desea obtner la información}

\item{fechaFinal}{Fecha hasta donde se desea obtener la informacion, puede
usar \code{\link[=Sys.Date]{Sys.Date()}} para obtener la informacion hasta la fecha actual}

\item{token}{Token proporcionado por Banxico para la consulta. Puede
obtenerlo en la pagina de
\href{https://www.banxico.org.mx/SieAPIRest/service/v1/token}{Banxico}}

\item{idioma}{\code{character}  Idioma para la consulta de los metadatos.
\itemize{
\item "es" Español
\item "en" Inglés
Por default la informacion se obtiene en español
}}
}
\value{
\code{\link[=list]{list()}} con los siguientes elementos:
\itemize{
\item \strong{Metadatos de la información de la tasa}
\item \strong{Data.Frame con la información procesada}
\item \strong{Informacion del mercado secundario}
}
}
\description{
Obtiene la informacion actualizada referente a un Bono Gubernamental de
México, esta informacion es btenida desde la API de Banxico y sitios de
informacion financiera.
}
\examples{
token <- '98f028f762387fd81728858fca4cc4e1ddaa4c538e3d6209f256a6a0ef25021b'
cetes28 <- getBono(bono = 'cetes28', fechaInicial = '2024-01-01',
fechaFinal = Sys.Date(), token = token)
}
