% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrap.R
\name{webscrapBonos}
\alias{webscrapBonos}
\title{WebscrapBonos}
\usage{
webscrapBonos()
}
\value{
\code{\link[=data.frame]{data.frame()}} con la informacion actualizada
}
\description{
Proporciona informacion en tiempo real del mercado secundario de bonos, la
informacion es obtenida de www.investing.com.
}
\examples{
webscrapBonos()
}
